requirements.txt
# Python 3.12+ required
aiofiles==23.2.1 # Added aiofiles
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
argcomplete==3.6.2
astroid==3.3.9
async-timeout==5.0.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.3.0
boto3==1.38.8
botocore==1.38.8
cachetools==5.3.3
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cohere==5.15.0
colorama==0.4.6
coverage==7.8.0
cryptography==44.0.3
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
flake8==7.0.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.42
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.23.3
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.7.3
groq==0.24.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isort==5.13.2
jiter==0.9.0
jmespath==1.0.1
kombu==5.5.3
logfire==3.14.1
logfire-api==3.14.1
markdown-it-py==3.0.0
mccabe==0.7.0
mcp==1.7.1
mdurl==0.1.2
mistralai==1.7.0
mypy==1.9.0
mypy_extensions==1.1.0
nodeenv==1.9.1
openai==1.77.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==25.0
pandas==2.2.2 # Added pandas
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.4
pydantic-ai==0.1.9
pydantic-ai-slim==0.1.9
pydantic-evals==0.1.9
pydantic-graph==0.1.9
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.2.0
PyGithub==2.3.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
PyNaCl==1.5.0
pyparsing==3.2.3
pyright==1.1.400
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==25.1.2
redis==6.0.0
requests==2.32.3
retry # Added retry library
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
setuptools==80.3.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.28
sse-starlette==2.3.3
starlette==0.46.2
tokenizers==0.21.1
tomlkit==0.13.2
tqdm==4.67.1
types-beautifulsoup4==4.12.0.20240229
types-html5lib==1.1.11.20241018
types-requests==2.31.0.20240311
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
